+++ src/screens/home/views/DiagnosedShareUploadView.tsx
@@ -1,9 +1,8 @@
 import React, {useCallback} from 'react';
 import {useI18n} from 'locale';
 import {useNavigation} from '@react-navigation/native';
-import {Text, ButtonSingleLine, Box} from 'components';
+import {Text, ButtonSingleLine, RoundedBox} from 'components';
 import {useAccessibilityAutoFocus} from 'shared/useAccessibilityAutoFocus';
-import {StyleSheet, Platform} from 'react-native';
 import AsyncStorage from '@react-native-community/async-storage';
 import {INITIAL_TEK_UPLOAD_COMPLETE} from 'shared/DataSharing';
 
@@ -21,37 +20,20 @@ export const DiagnosedShareUploadView = ({isBottomSheetExpanded}: {isBottomSheet
 
   return (
     <BaseHomeView iconName="hand-reminder-red">
-      <Box style={styles.roundedBox} backgroundColor="bodyTitleWhite">
-        <Box paddingHorizontal="m" paddingVertical="m">
-          <Text
-            focusRef={autoFocusRef}
-            variant="bodyTitle"
-            color="bodyText"
-            marginBottom="m"
-            accessibilityRole="header"
-          >
-            {i18n.translate('Home.DiagnosedShareUploadView.Title')}
-          </Text>
-          <Text variant="bodyText" color="bodyText" marginBottom="m">
-            {i18n.translate('Home.DiagnosedShareUploadView.Body1')}
-          </Text>
-          <ButtonSingleLine
-            text={i18n.translate('Home.DiagnosedShareUploadView.ButtonCTA')}
-            variant="dangerWhiteText"
-            onPress={toDataShare}
-            iconName="icon-chevron-white"
-          />
-        </Box>
-      </Box>
+      <RoundedBox>
+        <Text focusRef={autoFocusRef} variant="bodyTitle" color="bodyText" marginBottom="m" accessibilityRole="header">
+          {i18n.translate('Home.DiagnosedShareUploadView.Title')}
+        </Text>
+        <Text variant="bodyText" color="bodyText" marginBottom="m">
+          {i18n.translate('Home.DiagnosedShareUploadView.Body1')}
+        </Text>
+        <ButtonSingleLine
+          text={i18n.translate('Home.DiagnosedShareUploadView.ButtonCTA')}
+          variant="dangerWhiteText"
+          onPress={toDataShare}
+          iconName="icon-chevron-white"
+        />
+      </RoundedBox>
     </BaseHomeView>
   );
 };
-
-const styles = StyleSheet.create({
-  roundedBox: {
-    marginTop: Platform.OS === 'ios' ? 5 : -20,
-    backgroundColor: 'white',
-    borderRadius: 10,
-    zIndex: -1,
-  },
-});
